# doc/api.yaml
openapi: 3.0.3
info: 
  title: WASAPhoto
  description: |
    This API document describes the functioning of WASAPhoto website
  version: 1.0.0
paths:
...
  /users/{id}/follow/{otherUserId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/OtherUserId"
    put:
      tags:
        - Follow
      security:
        - bearerAuth: []
      operationId: followUser
      summary: follows another user
      description: allows the authorised user to follow another user
      responses:
        "204": { description: user followed successfully }
        "404": { description: user not found }
        "401": {description: operation unauthorized }
        "500": { description: failed to follow this user }
    delete:
      tags: 
        - Follow
      security:
        - bearerAuth: []
      operationId: unfollowUser
      summary: unfollows another user
      description: allows user to unfollow another user
      responses:
        "204": { description: user unfollowed successfully }
        "404": { description: user not found }
        "401": {description: operation unauthorized }
        "500": { description: failed to unfollow this user }
  /users/{id}/banned/{otherUserId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/OtherUserId"
    put:
      tags:
        - Ban 
      security:
        - bearerAuth: []
      operationId: banUser
      summary: bans user 
      description: allows user to ban others
      responses:
        "204": { description: user banned successfully }
        "404": { description: user not found }
        "401": {description: operation unauthorized }
        "500": { description: failed to ban user }
    delete:
      tags: 
        - Ban
      security:
        - bearerAuth: []
      operationId: unbanUser
      summary: unbans user
      description: allows user to remove a ban put previously
      responses:
        "204": { description: user unbanned successfully }
        "404": { description: error; user not found }
        "401": {description: operation unauthorized }
        "500": { description: failed to unban user }
  /users/{id}/photos/:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: 
        - Photos
      security:
        - bearerAuth: []
      operationId: uploadPhoto
      summary: uploads a new photo in the user profile
      description: |
        method that enables user to add a new photo to his profile
        generating a photo id for the photo
      requestBody:
        content:
          image/png:
            schema:
              $ref: "#/components/schemas/RawPhoto"
        required: true
      responses:
        "201":
          description: photo uploaded successfully
          content:
            image/png: 
              schema: { $ref: "#/components/schemas/Photo" }
        "404": {description: photo not found }
        "401": {description: operation unauthorized }
        "500": { description: an error occured please try again later }
...
  /users/{id}/photos/{photoId}/likes/{otherUserId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/PhotoId"
      - $ref: "#/components/parameters/OtherUserId"
    put:
      tags: 
        - Likes
      security:
        - bearerAuth: []
      operationId: likePhoto
      summary: to like a photo
      description: lets the user like a photo 
      responses:
        "204": 
          description: photo liked successfully
        "404": { description: photo not found }
        "401": {description: operation unauthorized }
        "500":  
          description: |
            an errror occured, please try again later 
    delete:
      tags: 
        - Likes
      security:
        - bearerAuth: []
      operationId: unlikePhoto
      summary: unlikes a photo
      description: allows the current user to remove a like from photo
      responses:
        "204": { description: like removed successfully }
        "404": { description: error; photo not found }
        "401": {description: operation unauthorized }
        "500": 
          description: |
            an error occured, please retry later 
...
  /users/{id}/photos/{photoId}/comments/{commentId}:
    parameters:
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/PhotoId"
      - $ref: "#/components/parameters/CommentId"
    delete:
      tags: 
        - Comments
      security:
        - bearerAuth: []
      operationId: uncommentPhoto
      summary: removes a comment
      description: allows user to delete his comment from a photo
      responses:
        "204": { description: comment removed successfully }
        "404": { description: photo not found }
        "401": {description: operation unauthorized }
        "500": 
          description: | 
            an error occured, please retry later 
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Username: 
      title: Username
      description: username of user 
      type: object
      properties:
        username:
          description: string between 5 and 15 characters
          type: string 
          pattern: '^([a-z0-9._-])\w+$'
          minLength: 5
          maxLength: 15
          example: maria77
    RawPhoto:
      description: contains the raw image the user wants to upload
      type: object
      properties:
        data: 
          description: data of the image
          type: string
          format: binary
          pattern: "^[0-1]*$"
          minLength: 1
          maxLength: 2147483647
          example: "0110010101"
    Photo:
      title: Photo
      description: | 
        this object represents a picture with all its 
        related information
      type: object
      properties:
        photoId:
          description: id of a photo
          type: number
          example: 10097
        owner:
          description: username of user that uploaded the photo
          $ref: "#/components/schemas/Username"
        likes:
          description: represents the lits of likes on a given photo
          type: array
          items:
            $ref: "#/components/schemas/Username"
          minItems: 0
          maxItems: 100000000000
          example: [maria77, bob77]
        comments: 
          description: represents the list of comments on a given photo
          type: array 
          items: 
            $ref: "#/components/schemas/Comment"
          minItems: 0
          maxItems: 100000000000
        dateAndTime:
          description: represents the date and time in which the picture was uploaded
          type: string
          format: date-time 
          minLength: 20
          maxLength: 20
          example: 2017-07-21T17:32:28Z 
    Comment: 
      title: Comment
      description: this object represents a single comment
      type: object
      properties:
        commentId:
          type: number
          description: represents the id of a comment 
          example: 01
        username: { $ref: "#/components/schemas/Username" }
        comment: 
          type: string 
          description: the comment itself 
          minLength: 1
          maxLength: 500 
          example: beautiful picture man! 
          pattern: "^(.*?)$"
...
  parameters:
    Id:
      schema:
        type: string
        pattern: '^([a-z0-9._-])\w+$'
        minLength: 5
        maxLength: 15
        example: abcdef012345
      name: id
      in: path
      required: true
      description: id of the authorised user
...
    OtherUserId:
      schema:
        type: string
        pattern: '^([a-z0-9._-])\w+$'
        minLength: 5
        maxLength: 15
        example: abcdef012345
      name: otherUserId
      in: path
      required: true
      description: id of a user
    PhotoId:
      schema:
        type: number
        example: 10097
      name: photoId
      in: path
      required: true
      description: id of a photo 
    CommentId:  
      schema:
        type: number
        example: 01
      name: commentId
      in: path
      required: true
      description: id of a comment
...